lotta stuff

+++ libft.h
@@ -6,26 +6,25 @@
 /*   By: yokitane <yokitane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/26 11:21:56 by yokitane          #+#    #+#             */
-/*   Updated: 2024/08/27 12:35:13 by yokitane         ###   ########.fr       */
+/*   Updated: 2024/08/29 14:15:10 by yokitane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-int	ft_isalpha(int c);
-int	ft_isdigit(int c);
-int	ft_isalnum(int c);
-int	ft_isascii(int c);
-int	ft_isprint(int c);
-int	ft_strlen(const char *str);
+# include	<stdlib.h>
+
+int		ft_isalpha(int c);
+int		ft_isdigit(int c);
+int		ft_isalnum(int c);
+int		ft_isascii(int c);
+int		ft_isprint(int c);
+size_t	ft_strlen(const char *str);
 void	*ft_memset(void *s, int c, int n);
-void	ft_bzero(void *s, int n);
-void	*ft_memcpy(void *dest, const void *src, int n);
-void	*ft_memmove(void *dest, const void *src, int n);
-int	ft_strlcpy(char *dst, const char *src, int size);
-int	ft_strlcat(char *dst, const char *src, int size);
-
-
-
+void	ft_bzero(void *s, size_t n);
+void	*ft_memcpy(void *dest, const void *src, size_t n);
+void	*ft_memmove(void *dest, const void *src, size_t n);
+size_t	ft_strlcpy(char *dst, const char *src, size_t size);
+size_t	ft_strlcat(char *dst, const char *src, size_t size);
 #endif
